EntityPinkBallProjectile
