+++ config/splat.us.stdre.yaml
@@ -37,7 +37,10 @@ segments:
       - [0x11770, rodata]
       - [0x117A0, rodata]
       - [0x117F0, rodata]
+      - [0x11808, .rodata, 14214]
+      - [0x1181C, rodata]
       - [0x118D0, rodata]
-      - [0x11A64, c]
+      - [0x11A64, c, 11A64]
+      - [0x14214, c, 14214]
       - [0x23260, data]
   - [0x23FCC]
